refactor to trim deps
