+++ .vscode/settings.json
@@ -1,6 +1,10 @@
 {
   "typescript.tsdk": "node_modules/typescript/lib",
   "cSpell.words": [
-    "Unextractable"
+    "bafyreiaxnmoptsqiehdff",
+    "blpptvdbenxcz",
+    "Unextractable",
+    "xgrbojw",
+    "xovn"
   ]
 }
