+++ packages/core/package.json
@@ -84,6 +84,10 @@
       "types": "./dist/src/cbor.d.ts",
       "import": "./src/cbor.js"
     },
+    "./car": {
+      "types": "./dist/src/car.d.ts",
+      "import": "./src/car.js"
+    },
     "./dag": {
       "types": "./dist/src/dag.d.ts",
       "import": "./src/dag.js"
