+++ packages/core/src/lib.js
@@ -4,6 +4,7 @@ export * as Invocation from './invocation.js'
 export * as Receipt from './receipt.js'
 export * as DAG from './dag.js'
 export * as CBOR from './cbor.js'
+export * as CAR from './car.js'
 export { delegate, isDelegation } from './delegation.js'
 export { invoke } from './invocation.js'
 export {
@@ -13,6 +14,7 @@ export {
   parse as parseLink,
   decode as decodeLink,
 } from './link.js'
+export { sha256 } from 'multiformats/hashes/sha2'
 export * as UCAN from '@ipld/dag-ucan'
 export * as DID from '@ipld/dag-ucan/did'
 export * as Signature from '@ipld/dag-ucan/signature'
