+++ packages/core/src/link.js
@@ -1,1 +1,3 @@
 export * from 'multiformats/link'
+export { base32 } from 'multiformats/bases/base32'
+export { base58btc } from 'multiformats/bases/base58'
