+++ packages/server/package.json
@@ -45,7 +45,6 @@
     "chai": "^4.3.6",
     "chai-subset": "^1.6.0",
     "mocha": "^10.1.0",
-    "multiformats": "^11.0.0",
     "nyc": "^15.1.0",
     "playwright-test": "^8.2.0",
     "typescript": "^4.9.5"
