+++ packages/server/test/handler.spec.js
@@ -2,7 +2,6 @@ import * as Client from '@ucanto/client'
 import * as Server from '../src/server.js'
 import * as Provider from '../src/handler.js'
 import * as CAR from '@ucanto/transport/car'
-import * as CBOR from '@ucanto/transport/cbor'
 import * as API from '@ucanto/interface'
 import { alice, bob, mallory, service } from './fixtures.js'
 import { test, assert } from './test.js'
