+++ packages/transport/src/lib.js
@@ -1,4 +1,3 @@
-export * as CBOR from './cbor.js'
 export * as CAR from './car.js'
 export * as JWT from './jwt.js'
 export * as HTTP from './http.js'
